change all class service to core, add new class for create bill

+++ cinema/cinema-core/src/main/java/com/nongviet201/cinema/core/repository/CinemaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CinemaRepository extends JpaRepository<Cinema, Long> {
+public interface CinemaRepository extends JpaRepository<Cinema, Integer> {
     Cinema findById(int id);
 
     List<Cinema> findAllByCity_Id(int cityId);
